move seeded context fixture to `conftest.py`
